+++ flutter/waitingnow/lib/src/Screen/Admin/waiting/EndAdminWidgetList.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:waitingnow/src/Get/WaitingAdminGet.dart';
-import 'package:waitingnow/src/Screen/Widget/WaitingAdminWidget.dart';
+import 'package:waitingnow/src/Screen/Widget/EndAdminWidget.dart';
 
-import '../../Controller/TimerController.dart';
+import '../../../Controller/TimerController.dart';
 
 class EndAdminWidgetList extends StatelessWidget {
   final TimerController timerController;
@@ -15,18 +15,16 @@ class EndAdminWidgetList extends StatelessWidget {
   Widget build(BuildContext context) {
     return Obx(() {
       List<dynamic> items = waitingAdminGet.endVO.value;
-      return Expanded(
-        child: ListView.builder(
+      return ListView.builder(
           itemCount: items.length,
           itemBuilder: (context, index) {
             return buildPage(items[index], index + 1);
           },
-        ),
-      );
+        );
     });
   }
 
   Widget buildPage(dynamic item, int index) {
-    return WaitingAdminWidget(item, index, timerController);
+    return EndAdminWidget(item, index);
   }
 }
