+++ flutter/waitingnow/lib/src/Screen/Widget/WaitingAdminWidget.dart
@@ -4,6 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:waitingnow/src/Screen/Admin/Admin.dart';
 import 'package:waitingnow/src/Controller/TimerController.dart';
+import 'package:waitingnow/src/Screen/Admin/Table/TableAssign.dart';
 
 import '../../Controller/WaitingController.dart';
 import '../../Domain/WaitingVO.dart';
@@ -311,7 +312,7 @@ class _WaitingAdminWidgetState extends State<WaitingAdminWidget> {
                   SizedBox(width: 25,),
                   ElevatedButton.icon(
                       onPressed: () {
-                        Get.offAll(Admin(), arguments: 1);
+                        Get.to(TableAssign());
                       },
                       icon: Icon(Icons.check),
                       style: ElevatedButton.styleFrom(
